power: Enable boostpulse on interaction hints
We have replaced some important CPU_BOOST hints, such as flings,
with the global fling detector introduced in Marshmallow.

Let's enable boosts for the INTERACTION hint to cover these important
UX cases. One side effect of this is that we also boost on user
interaction, which happens to be quite nice too.

Change-Id: Icb21b2bd9aebbed2ee934c7ea0abbf2854dd74f6